Router test passing with res(200)
