+++ package.json
@@ -35,6 +35,7 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "node-sass": "^9.0.0",
+    "regenerator-runtime": "^0.14.0",
     "sass": "^1.69.4",
     "sass-loader": "^13.3.2",
     "style-loader": "^3.3.3",
@@ -66,7 +67,8 @@
   "homepage": "https://github.com/Shape-Quest/Shape-Quest#reame",
   "jest": {
     "testEnvironment": "node",
-    "setupFiles": ["<rootDir>/jest.setup.js"]
-
+    "setupFiles": [
+      "<rootDir>/jest.setup.js"
+    ]
   }
 }
